pyproject.toml
[project]
name = "agentsystems-sdk"
version = "0.5.2"
description = "AgentSystems Python SDK and CLI"
readme = "README.md"
authors = [ { name = "AgentSystems", email = "support@agentsystems.ai" } ]
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = ["typer[all]>=0.9", "python-dotenv>=1.0", "rich>=13.7", "docker>=7.0", "PyYAML>=6.0", "requests>=2.31"]

license-files = ["LICENSE", "NOTICE"]

[project.scripts]
agentsystems = "agentsystems_sdk.cli:app"
agntsys = "agentsystems_sdk.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentsystems_sdk*"]

[tool.setuptools.package-data]
agentsystems_sdk = ["deployments_scaffold/**/*", "licenses/**/*"]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"
